HPE_Mentor/JobTrainer/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
SQLAlchemy==2.0.0
python-dotenv == 1.0.0
psycopg2 == 2.9.7
python-multipart == 0.0.6